refactor: main dispatches, each rpc method is a function, and the daemon state takes its config
