+++ bansheed/src/api/tests.rs
@@ -595,16 +595,11 @@ async fn speak_passes_the_voice_parameter_to_the_backend() {
     let speech =
         crate::text_to_speech::SpeechPlayer::new(Box::new(VoiceCapture(Arc::clone(&captured))));
     let state = Arc::new(DaemonState::new(
-        "0.0.0",
-        "stt",
-        "vad",
-        0.5,
-        "default".to_string(),
+        Arc::new(crate::config::Config::default()),
         None,
         speech,
         std::sync::mpsc::channel().0,
         crate::audio::cues::Cues::silent(),
-        crate::config::BargeInMode::Stop,
     ));
 
     let response = dispatch(
