+++ bansheed/src/test_support.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use crate::config::BargeInMode;
+use crate::config::Config;
 use crate::history::TranscriptionHistory;
 use crate::state::{ConsumerCommand, DaemonState};
 use crate::text_to_speech::{ActiveUtterance, SpeechPlayer, TtsBackend};
@@ -58,18 +58,35 @@ fn state(
     speech: SpeechPlayer,
     commands: std::sync::mpsc::Sender<ConsumerCommand>,
 ) -> Arc<DaemonState> {
+    state_running(Config::default(), history, speech, commands)
+}
+
+fn state_running(
+    config: Config,
+    history: Option<rusqlite::Connection>,
+    speech: SpeechPlayer,
+    commands: std::sync::mpsc::Sender<ConsumerCommand>,
+) -> Arc<DaemonState> {
     Arc::new(DaemonState::new(
-        "0.0.0",
-        "stt",
-        "vad",
-        0.5,
-        "default".to_string(),
+        Arc::new(config),
         history,
         speech,
         commands,
         crate::audio::cues::Cues::silent(),
-        BargeInMode::Stop,
     ))
+}
+
+/// A daemon state that started with `config`, so its restart-only keys run those values.
+pub fn daemon_state_running(
+    config: Config,
+    commands: std::sync::mpsc::Sender<ConsumerCommand>,
+) -> Arc<DaemonState> {
+    state_running(
+        config,
+        None,
+        SpeechPlayer::new(Box::new(NullBackend)),
+        commands,
+    )
 }
 
 /// Pass a real sender only when the test reads the command receiver.
